refactor(State): state에서 List<Card> -> Hand를 반환하는 메서드로 변경
