+++ src/main/java/blackjack/model/state/State.java
@@ -2,14 +2,14 @@ package blackjack.model.state;
 
 import blackjack.model.Money;
 import blackjack.model.deck.Card;
-import java.util.List;
+import blackjack.model.participant.Hand;
 
 public interface State {
     State draw(Card card);
 
     State stand();
 
-    List<Card> getCards();
+    Hand hand();
 
     int calculateScore();
