+++ src/main/java/blackjack/model/state/finished/BlackJack.java
@@ -1,30 +1,30 @@
-package blackjack.model.state;
+package blackjack.model.state.finished;
 
 import blackjack.model.Money;
 import blackjack.model.deck.Card;
 import blackjack.model.participant.Hand;
-import java.util.List;
+import blackjack.model.state.State;
 
-public class Bust implements State {
+public class BlackJack implements State {
     private final Hand hand;
 
-    public Bust(final Hand hand) {
+    public BlackJack(final Hand hand) {
         this.hand = hand;
     }
 
     @Override
     public State draw(final Card card) {
-        throw new UnsupportedOperationException("버스트 상태에서는 카드를 더 받을 수 없습니다.");
+        throw new UnsupportedOperationException("추가로 카드를 받을 수 없습니다.");
     }
 
     @Override
     public State stand() {
-        throw new UnsupportedOperationException("이미 버스트이므로 카드를 받지 않을지 여부를 선택할 수 없습니다.");
+        throw new UnsupportedOperationException("이미 블랙잭이므로 카드를 받지 않을지 여부를 선택할 수 없습니다.");
     }
 
     @Override
-    public List<Card> getCards() {
-        return hand.getCards();
+    public Hand hand() {
+        return hand;
     }
 
     @Override
